+++ r2-lcp/src/main/java/org/readium/r2/lcp/service/LicensesService.kt
@@ -19,11 +19,13 @@ import org.readium.r2.lcp.license.container.BytesLicenseContainer
 import org.readium.r2.lcp.license.container.LicenseContainer
 import org.readium.r2.lcp.license.container.createLicenseContainer
 import org.readium.r2.lcp.license.model.LicenseDocument
+import org.readium.r2.shared.format.Format
 import java.io.File
 import org.readium.r2.shared.util.Try
 import timber.log.Timber
+import java.util.Properties
+import java.util.UUID
 import kotlin.coroutines.resume
-import kotlin.coroutines.suspendCoroutine
 
 
 internal class LicensesService(
@@ -35,22 +37,33 @@ internal class LicensesService(
     private val context: Context
 ) : LcpService, LCPService, CoroutineScope by MainScope() {
 
-    override suspend fun importPublication(lcpl: ByteArray, authentication: LcpAuthenticating?, allowUserInteraction: Boolean, sender: Any?): Try<LcpService.ImportedPublication, LcpException>? =
+    override suspend fun importPublication(lcpl: ByteArray): Try<LcpService.ImportedPublication, LcpException> =
         try {
-            val container = BytesLicenseContainer(lcpl)
-            retrieveLicense(container, authentication, allowUserInteraction, sender)
-                ?.fetchPublication(context)
-                ?.let { Try.success(it) }
+            val licenseDocument = LicenseDocument(lcpl)
+            if (DEBUG) Timber.d("license ${licenseDocument.json}")
+            fetchPublication(licenseDocument, context).let { Try.success(it) }
         } catch (e:Exception) {
            Try.failure(LcpException.wrap(e))
         }
 
     override fun importPublication(lcpl: ByteArray, authentication: LCPAuthenticating?, completion: (LCPImportedPublication?, LCPError?) -> Unit) {
-        launch {
-            importPublication(lcpl, authentication.toLcpAuthenticating(), authentication != null, null)
-                ?.onSuccess { completion(it.toLCPImportedPublication(), null) }
-                ?.onFailure { completion(null, it.toLCPError()) }
-                ?: run { completion(null, null) }
+        val container = BytesLicenseContainer(lcpl)
+        try {
+            retrieveLicense(container, authentication.toLcpAuthenticating(), allowUserInteraction = true, sender = null) { license ->
+                if (license != null) {
+                    launch {
+                        try {
+                            completion(fetchPublication(license.license, context).toLCPImportedPublication(), null)
+                        } catch (e: Exception) {
+                            completion(null, LCPError.wrap(e))
+                        }
+                    }
+                } else {
+                    completion(null, null)
+                }
+            }
+        } catch (e:Exception) {
+            completion(null, LCPError.wrap(e))
         }
     }
 
@@ -72,7 +85,7 @@ internal class LicensesService(
     override fun retrieveLicense(publication: String, authentication: LCPAuthenticating?, completion: (LCPLicense?, LCPError?) -> Unit) {
         try {
             val container = runBlocking { createLicenseContainer(publication) }
-            retrieveLicense(container, authentication.toLcpAuthenticating(), authentication != null, null) { license ->
+            retrieveLicense(container, authentication.toLcpAuthenticating(), true, null) { license ->
                 Timber.d("license retrieved ${license?.license}")
                 completion(license, null)
             }
@@ -134,6 +147,39 @@ internal class LicensesService(
                 completion(null)
             }
         }
+    }
+
+    internal suspend fun fetchPublication(license: LicenseDocument, context: Context): LcpService.ImportedPublication {
+        val link = license.link(LicenseDocument.Rel.publication)
+        val url = link?.url
+            ?: throw LcpException.Parsing.Url(rel = LicenseDocument.Rel.publication.rawValue)
+
+        val properties =  Properties()
+        withContext(Dispatchers.IO) {
+            context.assets.open("configs/config.properties").let { properties.load(it) }
+        }
+        val useExternalFileDir = properties.getProperty("useExternalFileDir", "false")!!.toBoolean()
+
+        val rootDir: String =  if (useExternalFileDir) {
+            context.getExternalFilesDir(null)?.path + "/"
+        } else {
+            context.filesDir.path + "/"
+        }
+
+        val fileName = UUID.randomUUID().toString()
+        val destination = File(rootDir, fileName)
+        if (DEBUG) Timber.i("LCP destination $destination")
+
+        val format = network.download(url, destination) ?: Format.of(mediaType = link.type) ?: Format.EPUB
+
+        // Saves the License Document into the downloaded publication
+        val container = createLicenseContainer(destination.path, format)
+        container.write(license)
+
+        return LcpService.ImportedPublication(
+            localURL = destination.path,
+            suggestedFilename = "${license.id}.${format.fileExtension}"
+        )
     }
 
 }
